requirements.txt
python-binance
click
python-dotenv